checkout: do not delete directories with untracked entries
If the `GIT_CHECKOUT_FORCE` flag is given to any of the `git_checkout`
invocations, we remove files which were previously staged. But while
doing so, we unfortunately also remove unstaged files in a directory
which contains at least one staged file, resulting in potential data
loss.

The issue resides in our logic which tries to determine which action to
excute for a tree entry. We currently assume that a directory is safe to
be deleted as soon as it contains at least one versioned entry - which
is ovbiously wrong. Instead, we first have to determine whether the
directory contains any unversioned files and, if so, advance into the
directory itself. As such, the directory will not be deleted, but all
versioned entries inside of the directory will be reset correctly.

This commit also adds two tests to verify behavior.