ssh: don't try to look up when linked at compile-time
Instead of trying to use the same functions by looking them up at startup, rely
on the linker to give us the right functions and simply define away the `git_`
prefix.

This should let us work with both compile-time and run-time linking, though it
introduces a lot of conditional compilation.