clar: introduce concept of known breakages
Currently, we have no way to record known breakages in test suites. If a
projects knows of a certain way in which functionality is broken but
cannot yet fix it, it may want to record this fact. This provides some
benefits:

- breakage is recorded and easily discoverable, so that committers may
  pick out known breakages to pick up work

- it becomes possible to submit non-working tests first demonstrating
  breakage and later on submit a fix, switching over to a normal assert

- an error can be thrown as soon as the test is _not_ broken anymore,
  clearly highlighting when another change fixes breakage

This commit introduces a new state `CL_TEST_BROKEN` to mark a test as
broken. It also introduces a new macro `cl_break` which calls into the
new function `clar__broken`, which can be used to mark a test case as
broken.