errors: introduce overflow error class
Due to their previous conotation of having to do with memory
allocations, the `GITERR_CHECK_ADD`/`GITERR_CHECK_MULTIPLY` macros
currently set an out-of-memory error on overflow. This is a rather
misleading error when the macros were used in a context which has
nothing to do with memory allocations but simply for overflow-sensitive
integer operations.

Introduce a new error class `GITERR_OVERFLOW` which is used by these
macros.